- Layout angepasst - Uebersetzungen hinzugefuegt - Kommentare hinzugefuegt

+++ src/main/java/com/springboot/vitalorganize/model/Meal.java
@@ -4,33 +4,33 @@ import java.time.LocalDate;
 
 public class Meal {
 
-    private int id;
-    private int userId;
-    private int recipeId;
+    private Long id;
+    private Long userId;
+    private Long recipeId;
     private LocalDate cookDate;
 
     // Getter und Setter
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public int getRecipeId() {
+    public Long getRecipeId() {
         return recipeId;
     }
 
-    public void setRecipeId(int recipeId) {
+    public void setRecipeId(Long recipeId) {
         this.recipeId = recipeId;
     }
